Moved GenerationError to error.rs
