+++ chain-signatures/node/src/protocol/mod.rs
@@ -2,6 +2,7 @@ mod cryptography;
 
 pub mod consensus;
 pub mod contract;
+pub mod error;
 pub mod message;
 pub mod presignature;
 pub mod signature;
