+++ chain-signatures/node/src/protocol/presignature.rs
@@ -1,8 +1,8 @@
 use super::message::{MessageChannel, PresignatureMessage};
-use super::signature::SignId;
 use super::state::RunningState;
 use super::triple::{Triple, TripleId, TripleManager};
 use crate::protocol::contract::primitives::Participants;
+use crate::protocol::error::GenerationError;
 use crate::storage::presignature_storage::PresignatureStorage;
 use crate::types::{PresignatureProtocol, SecretKeyShare};
 use crate::util::AffinePointExt;
@@ -125,36 +125,6 @@ impl PresignatureGenerator {
 
         self.protocol.poke()
     }
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum GenerationError {
-    #[error("presignature already generated")]
-    AlreadyGenerated,
-    #[error("cait-sith initialization error: {0}")]
-    CaitSithInitializationError(#[from] InitializationError),
-    #[error("triple storage error: {0}")]
-    TripleStoreError(String),
-    #[error("triple {0} is generating")]
-    TripleIsGenerating(TripleId),
-    #[error("triple {0} is in garbage collection")]
-    TripleIsGarbageCollected(TripleId),
-    #[error("triple access denied: id={0}, {1}")]
-    TripleDenied(TripleId, &'static str),
-    #[error("presignature {0} is generating")]
-    PresignatureIsGenerating(PresignatureId),
-    #[error("presignature {0} is missing")]
-    PresignatureIsMissing(PresignatureId),
-    #[error("presignature {0} is in garbage collection")]
-    PresignatureIsGarbageCollected(TripleId),
-    #[error("presignature access denied: id={0}, {1}")]
-    PresignatureDenied(PresignatureId, &'static str),
-    #[error("presignature bad parameters")]
-    PresignatureBadParameters,
-    #[error("waiting for missing sign request id={0:?}")]
-    WaitingForIndexer(SignId),
-    #[error("invalid proposer expected={0:?}, actual={1:?}")]
-    InvalidProposer(Participant, Participant),
 }
 
 /// Abstracts how triples are generated by providing a way to request a new triple that will be
